[DONE][AuthController] Two-factor authentication

+++ app/Http/Controllers/DeveloperController.php
@@ -4,20 +4,36 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\DeveloperFormRequest;
 use App\Models\Developer;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Database\QueryException;
+use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\View\View;
 
 class DeveloperController extends Controller
 {
     public function index(): View
     {
-        $developers = Developer::orderBy('id', 'desc')->paginate(5);
+        try {
+            $developers = Developer::orderBy('id', 'desc')->paginate(5);
 
-        return view('pages.developers.index', compact('developers'));
+            return view('pages.developers.index', compact('developers'));
+        } catch (QueryException $exception) {
+            return back()->withError('An error occurred while loading developers.');
+        }
     }
 
-    public function show(Developer $developer): View
+    public function show($id): View
     {
-        return view('pages.developers.show', compact('developer'));
+        try {
+            $developer = Developer::findOrFail($id);
+
+            return view('pages.developers.show', compact('developer'));
+        } catch (ModelNotFoundException $e) {
+            return back()->withError('Developer not found.');
+        } catch (\Exception $e) {
+            return back()->withError('An error occurred while showing the developer.');
+        }
     }
 
     public function create(): View
@@ -25,29 +41,69 @@ class DeveloperController extends Controller
         return view('pages.developers.create');
     }
 
-    public function store(DeveloperFormRequest $request)
+    public function store(DeveloperFormRequest $request): RedirectResponse
     {
-        Developer::create($request->validated());
+        try {
+            $this->authorize('create', Developer::class);
 
-        return redirect()->route('developers.index');
+            $validatedData            = $request->validated();
+            $validatedData['user_id'] = Auth::id();
+
+            Developer::create($validatedData);
+
+            return redirect()->route('developers.index')->with('success', 'Developer created successfully!');
+        } catch (QueryException $exception) {
+            return back()->withError('An error occurred while storing developer.');
+        } catch (\Exception $e) {
+            return back()->withError('An error occurred while storing developer.');
+        }
     }
 
-    public function edit(Developer $developer): View
+    public function edit($id)
     {
-        return view('pages.developers.edit', compact('developer'));
+        try {
+            $developer = Developer::findOrFail($id);
+
+            $this->authorize('update', $developer);
+
+            return view('pages.developers.edit', compact('developer'));
+        } catch (ModelNotFoundException $exception) {
+            return back()->withError('Developer not found.');
+        } catch (\Exception $e) {
+            return back()->withError('An error occurred while editing developer.');
+        }
     }
 
-    public function update(DeveloperFormRequest $request, Developer $developer)
+    public function update(DeveloperFormRequest $request, Developer $developer): RedirectResponse
     {
-        $developer->update($request->validated());
+        try {
+            $this->authorize('update', $developer);
 
-        return redirect()->route('developers.index');
+            $developer->update($request->validated());
+
+            return redirect()->route('developers.index')->with('success', 'Developer updated successfully!');
+        } catch (QueryException $exception) {
+            return back()->withError('An error occurred while updating developer.');
+        } catch (\Exception $e) {
+            return back()->withError('An error occurred while updating developer.');
+        }
     }
 
-    public function destroy(Developer $developer)
+    public function destroy($id): RedirectResponse
     {
-        $developer->delete();
+        try {
+            $developer = Developer::findOrFail($id);
 
-        return redirect()->route('developers.index');
+            $this->authorize('delete', $developer);
+
+            $developer->delete();
+
+            return back()->with('success', 'Developer deleted successfully!');
+        } catch (ModelNotFoundException $exception) {
+            return back()->withError('Developer not found.');
+        } catch (\Exception $e) {
+            return back()->withError('An error occurred while deleting developer.');
+        }
     }
+
 }
